# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Client Portal API for a freelance design business
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ProjectStatus = (typeof ProjectStatus)[keyof typeof ProjectStatus];

export const ProjectStatus = {
  pending: "pending",
  in_progress: "in_progress",
  review: "review",
  completed: "completed",
  cancelled: "cancelled",
} as const;

/**
 * @nullable
 */
export type ProjectPriority =
  | (typeof ProjectPriority)[keyof typeof ProjectPriority]
  | null;

export const ProjectPriority = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface Project {
  id: number;
  clientName: string;
  /** @nullable */
  clientEmail?: string | null;
  title: string;
  /** @nullable */
  description?: string | null;
  status: ProjectStatus;
  /** @nullable */
  priority?: ProjectPriority;
  /** @nullable */
  dueDate?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type ProjectInputStatus =
  (typeof ProjectInputStatus)[keyof typeof ProjectInputStatus];

export const ProjectInputStatus = {
  pending: "pending",
  in_progress: "in_progress",
  review: "review",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export type ProjectInputPriority =
  (typeof ProjectInputPriority)[keyof typeof ProjectInputPriority];

export const ProjectInputPriority = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface ProjectInput {
  /** @minLength 1 */
  clientName: string;
  clientEmail?: string;
  /** @minLength 1 */
  title: string;
  description?: string;
  status: ProjectInputStatus;
  priority?: ProjectInputPriority;
  dueDate?: string;
}

export type ProjectUpdateStatus =
  (typeof ProjectUpdateStatus)[keyof typeof ProjectUpdateStatus];

export const ProjectUpdateStatus = {
  pending: "pending",
  in_progress: "in_progress",
  review: "review",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export type ProjectUpdatePriority =
  (typeof ProjectUpdatePriority)[keyof typeof ProjectUpdatePriority];

export const ProjectUpdatePriority = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface ProjectUpdate {
  clientName?: string;
  clientEmail?: string;
  title?: string;
  description?: string;
  status?: ProjectUpdateStatus;
  priority?: ProjectUpdatePriority;
  dueDate?: string;
}

export type DesignRequestStatus =
  (typeof DesignRequestStatus)[keyof typeof DesignRequestStatus];

export const DesignRequestStatus = {
  new: "new",
  reviewing: "reviewing",
  accepted: "accepted",
  rejected: "rejected",
} as const;

export interface DesignRequest {
  id: number;
  clientName: string;
  /** @nullable */
  clientEmail?: string | null;
  title: string;
  description: string;
  /** @nullable */
  requestType?: string | null;
  status: DesignRequestStatus;
  /** @nullable */
  projectId?: number | null;
  createdAt: string;
}

export interface DesignRequestInput {
  /** @minLength 1 */
  clientName: string;
  clientEmail?: string;
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  description: string;
  requestType?: string;
}

export type DesignRequestUpdateStatus =
  (typeof DesignRequestUpdateStatus)[keyof typeof DesignRequestUpdateStatus];

export const DesignRequestUpdateStatus = {
  new: "new",
  reviewing: "reviewing",
  accepted: "accepted",
  rejected: "rejected",
} as const;

export interface DesignRequestUpdate {
  status?: DesignRequestUpdateStatus;
  projectId?: number;
}

export interface ProjectFile {
  id: number;
  projectId: number;
  fileName: string;
  fileUrl: string;
  /** @nullable */
  fileSize?: number | null;
  /** @nullable */
  mimeType?: string | null;
  /** @nullable */
  uploadedBy?: string | null;
  uploadedAt: string;
}

export interface ProjectFileInput {
  fileName: string;
  fileUrl: string;
  fileSize?: number;
  mimeType?: string;
  uploadedBy?: string;
}

export type MessageRole = (typeof MessageRole)[keyof typeof MessageRole];

export const MessageRole = {
  client: "client",
  designer: "designer",
} as const;

export interface Message {
  id: number;
  projectId: number;
  author: string;
  role?: MessageRole;
  content: string;
  createdAt: string;
}

export type MessageInputRole =
  (typeof MessageInputRole)[keyof typeof MessageInputRole];

export const MessageInputRole = {
  client: "client",
  designer: "designer",
} as const;

export interface MessageInput {
  /** @minLength 1 */
  author: string;
  role?: MessageInputRole;
  /** @minLength 1 */
  content: string;
}

export interface DashboardSummary {
  totalProjects: number;
  activeProjects: number;
  pendingRequests: number;
  completedProjects: number;
  newRequestsThisWeek: number;
}

export type ActivityItemType =
  (typeof ActivityItemType)[keyof typeof ActivityItemType];

export const ActivityItemType = {
  project_created: "project_created",
  project_updated: "project_updated",
  request_submitted: "request_submitted",
  file_uploaded: "file_uploaded",
  message_sent: "message_sent",
} as const;

export interface ActivityItem {
  id: string;
  type: ActivityItemType;
  description: string;
  /** @nullable */
  projectId?: number | null;
  /** @nullable */
  projectTitle?: string | null;
  createdAt: string;
}

export interface StatusCount {
  status: string;
  count: number;
}

export type ListProjectsParams = {
  status?: string;
};

export type ListRequestsParams = {
  status?: string;
};