# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Client Portal API for a freelance design business
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: projects
    description: Design project management
  - name: requests
    description: Client design requests
  - name: files
    description: File uploads
  - name: messages
    description: Project messages/comments
  - name: dashboard
    description: Dashboard aggregates

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"

  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /requests:
    get:
      operationId: listRequests
      tags: [requests]
      summary: List all design requests
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DesignRequest"
    post:
      operationId: createRequest
      tags: [requests]
      summary: Submit a new design request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DesignRequestInput"
      responses:
        "201":
          description: Created request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DesignRequest"

  /requests/{id}:
    get:
      operationId: getRequest
      tags: [requests]
      summary: Get a design request by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Design request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DesignRequest"
        "404":
          description: Not found
    patch:
      operationId: updateRequest
      tags: [requests]
      summary: Update a design request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DesignRequestUpdate"
      responses:
        "200":
          description: Updated request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DesignRequest"
        "404":
          description: Not found
    delete:
      operationId: deleteRequest
      tags: [requests]
      summary: Delete a design request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /projects/{id}/files:
    get:
      operationId: listProjectFiles
      tags: [files]
      summary: List files for a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of files
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectFile"
    post:
      operationId: addProjectFile
      tags: [files]
      summary: Add a file record to a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectFileInput"
      responses:
        "201":
          description: Created file record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectFile"

  /projects/{id}/files/{fileId}:
    delete:
      operationId: deleteProjectFile
      tags: [files]
      summary: Remove a file from a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: fileId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /projects/{id}/messages:
    get:
      operationId: listProjectMessages
      tags: [messages]
      summary: List messages for a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: createProjectMessage
      tags: [messages]
      summary: Post a message on a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Created message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"

  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get dashboard summary stats
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"

  /dashboard/activity:
    get:
      operationId: getRecentActivity
      tags: [dashboard]
      summary: Get recent activity feed
      responses:
        "200":
          description: Recent activity items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"

  /projects/by-status:
    get:
      operationId: getProjectsByStatus
      tags: [projects]
      summary: Get project counts grouped by status
      responses:
        "200":
          description: Project counts by status
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StatusCount"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Project:
      type: object
      required: [id, clientName, title, status, createdAt, updatedAt]
      properties:
        id:
          type: integer
        clientName:
          type: string
        clientEmail:
          type: ["string", "null"]
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, in_progress, review, completed, cancelled]
        priority:
          type: ["string", "null"]
          enum: [low, medium, high, null]
        dueDate:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string

    ProjectInput:
      type: object
      required: [clientName, title, status]
      properties:
        clientName:
          type: string
          minLength: 1
        clientEmail:
          type: string
        title:
          type: string
          minLength: 1
        description:
          type: string
        status:
          type: string
          enum: [pending, in_progress, review, completed, cancelled]
        priority:
          type: string
          enum: [low, medium, high]
        dueDate:
          type: string

    ProjectUpdate:
      type: object
      properties:
        clientName:
          type: string
        clientEmail:
          type: string
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [pending, in_progress, review, completed, cancelled]
        priority:
          type: string
          enum: [low, medium, high]
        dueDate:
          type: string

    DesignRequest:
      type: object
      required: [id, clientName, title, description, status, createdAt]
      properties:
        id:
          type: integer
        clientName:
          type: string
        clientEmail:
          type: ["string", "null"]
        title:
          type: string
        description:
          type: string
        requestType:
          type: ["string", "null"]
        status:
          type: string
          enum: [new, reviewing, accepted, rejected]
        projectId:
          type: ["integer", "null"]
        createdAt:
          type: string

    DesignRequestInput:
      type: object
      required: [clientName, title, description]
      properties:
        clientName:
          type: string
          minLength: 1
        clientEmail:
          type: string
        title:
          type: string
          minLength: 1
        description:
          type: string
          minLength: 1
        requestType:
          type: string

    DesignRequestUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [new, reviewing, accepted, rejected]
        projectId:
          type: integer

    ProjectFile:
      type: object
      required: [id, projectId, fileName, fileUrl, uploadedAt]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        fileName:
          type: string
        fileUrl:
          type: string
        fileSize:
          type: ["integer", "null"]
        mimeType:
          type: ["string", "null"]
        uploadedBy:
          type: ["string", "null"]
        uploadedAt:
          type: string

    ProjectFileInput:
      type: object
      required: [fileName, fileUrl]
      properties:
        fileName:
          type: string
        fileUrl:
          type: string
        fileSize:
          type: integer
        mimeType:
          type: string
        uploadedBy:
          type: string

    Message:
      type: object
      required: [id, projectId, author, content, createdAt]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        author:
          type: string
        role:
          type: string
          enum: [client, designer]
        content:
          type: string
        createdAt:
          type: string

    MessageInput:
      type: object
      required: [author, content]
      properties:
        author:
          type: string
          minLength: 1
        role:
          type: string
          enum: [client, designer]
        content:
          type: string
          minLength: 1

    DashboardSummary:
      type: object
      required: [totalProjects, activeProjects, pendingRequests, completedProjects, newRequestsThisWeek]
      properties:
        totalProjects:
          type: integer
        activeProjects:
          type: integer
        pendingRequests:
          type: integer
        completedProjects:
          type: integer
        newRequestsThisWeek:
          type: integer

    ActivityItem:
      type: object
      required: [id, type, description, createdAt]
      properties:
        id:
          type: string
        type:
          type: string
          enum: [project_created, project_updated, request_submitted, file_uploaded, message_sent]
        description:
          type: string
        projectId:
          type: ["integer", "null"]
        projectTitle:
          type: ["string", "null"]
        createdAt:
          type: string

    StatusCount:
      type: object
      required: [status, count]
      properties:
        status:
          type: string
        count:
          type: integer