# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Client Portal API for a freelance design business
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  totalProjects: number;
  activeProjects: number;
  pendingRequests: number;
  completedProjects: number;
  newRequestsThisWeek: number;
}