# lib/api-zod/src/generated/types/designRequestInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Client Portal API for a freelance design business
 * OpenAPI spec version: 0.1.0
 */

export interface DesignRequestInput {
  /** @minLength 1 */
  clientName: string;
  clientEmail?: string;
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  description: string;
  requestType?: string;
}