# lib/api-zod/src/generated/types/designRequestUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Client Portal API for a freelance design business
 * OpenAPI spec version: 0.1.0
 */
import type { DesignRequestUpdateStatus } from "./designRequestUpdateStatus";

export interface DesignRequestUpdate {
  status?: DesignRequestUpdateStatus;
  projectId?: number;
}