# lib/api-zod/src/generated/types/messageInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Client Portal API for a freelance design business
 * OpenAPI spec version: 0.1.0
 */
import type { MessageInputRole } from "./messageInputRole";

export interface MessageInput {
  /** @minLength 1 */
  author: string;
  role?: MessageInputRole;
  /** @minLength 1 */
  content: string;
}